Update from original (#1)
* getResampledByCount does indeed add points if necessary

I tried calling getResampledByCount(30) on a polyline with 18 points, and the resulting line has 30 points, which seems to contradict what the documentation says.

* Update introduction.markdown

* Update ofFbo.markdown

* setPixelFormat description added

Description on setPixelFormat, it was unclear which ofPixelFormats were supported. In ofAVFoundationPlayer I found only these two ofPixelFormats are supported.

* "## 참여하기" 쪽 약간의 오타수정및 의역

* Revert ""## 참여하기" 쪽 약간의 오타수정및 의역"

* removed manual hyphenation

The word option was hyphenated in the source.

* Update How_to_load_and_display_an_image.markdown

* Old android studio tip that got lost in a PR

* Update slack links to heroku app to get around invitation

* setPixelFormat: specified OSX behaviour

Linux can support all the formats in the OF_PIXELS_FORMAT enum, but for OSX it's different.

* Fix broken feeds

* Add proxies for rss feeds to solve CORS problems

* fix auto build script

* Fix problem with code blocks

* code blocks 100% width

* fix code blocks stylesheet path

* fix missing scope resolution operator `::`

* Add ofTimer example

Taken from https://forum.openframeworks.cc/t/running-a-function-every-250us-microseconds-on-ras-pi-2/21250/2

* Now VS 2017 is out link to 2015 needs update

* Update for VS link 2015

* Update for VS link 2015

* updating to v0.9.8 + `make RunRelease` instead of `make run`

* Update Raspberry-Pi-Getting-Started.md

* vs setup guide: updated vs2015 download link

* Update about.md

* Minor text update

Changes the text to match the IDE used.

* Add missing startThread() to ofTimer example

* Typo from icoSphere cut and paste

* updated vs.md to prevent exception error

If you follow the install guide for Visual Studio users on the website, you'll run into an error when you want to create an OF project, because it didn't specify, that Common Tools for Visual C++ 2015 have to be installed. The problem is documented in the forums, the website just didn't reflect it yet. Hope this fixes it.

* 번역 시작

* Update ofVideoPlayer.markdown

* use update method instead of reloadTexture

* colorTwo is blue as the images

* ofxGui: mention ofParameter, add a brief description

* Fix documentation for ofFbo.end() function

* Update example code to remove deprecated function call

ofThread::startThread(bool mutexBlocks, bool verbose) is deprecated

* Change height of nightlybuild iframe

Android is cut off..

* Update introduction.markdown

* msys korean updated

* Fix documentation of ofVec3f::ofVec3f(const ofVec4f &vec)

* import_release: vs -> vs2015 & vs2017

* Change height of nightlybuilds iframe to accomodate vs2015 & vs2017

* import_release: linux64 -> linux64gcc4